🦄 refactor(Client|Notes): out of App.tsx
