+++ client/src/Services/note.service.ts
@@ -1,14 +1,14 @@
 import { fetchRequest } from 'Services';
-import { Note } from 'Types/index';
+import { Note } from 'Types';
 
 const NOTES_URL = '/notes';
 
-export function addNote (uid: string, note: Note): Promise<Note> {
+export function addNote (uid: string, note: string): Promise<Note> {
   return fetchRequest(`${NOTES_URL}/${uid}`, {
     method: 'POST',
     mode: 'cors',
     headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify(note)
+    body: JSON.stringify({note})
   })
 }
