+++ client/src/Types/Note.type.ts
@@ -1,5 +1,5 @@
 export type Note = {
-  id?: number
+  id: number
   note: string
   createdAt: Date
   uid: string
